Started a test component for Cards. This will be used in Card like Layouts.
- Adds the SFCard component.

The SFCard component is useful for creating a controlled layout for UI that needs a title header, body for content and a footer for extra things.

Example 1: Could be a Task Card for an Editor that is used for Task Management in Unity or for linking Task from Git Hub and visually representing them inside of Unity.

Example 2: Could be a portrait card for showing an image of a character, the name as the header title, than the body could be a stat list or a description of the character. Shatter Fantasy studio uses this for the custom RPG Editor tools for their Metroidvania projects.