Created a DragAndDrop Maniplulator to help create DragAndDrop functionality for UI elements.
Important Note: Another dev a long time ago shared this knowledge to help demonstrate this simple way to drag and drop. We will be adding his name in the credits file as a thank you for sharing his knowledge and insight.

The DragAndDrop Maniuplator allows UI elemnts to be draggable while handling pointer events. This can be used for a lot of thing.

Example 1: Dragging graph nodes in a graph editor. This could be a dialogue editor, AI editor, or a white boarding tool for brain storming.

Example2: Dragging an UIElement from container to container. Think of like a column of tasks where you can drag the task note to a different section to organize it.