Created a MouseDragRectManipulator for making Rect structs by dragging.
This one will have a dedicated documentation page in the future.

Added: MouseRectDragManipulator class
This class is a used for allowing mouse dragging to be read and see the delta or change in mouse position on release to create a Rect struct that can be used in different editor tools.

Example this Manipulator is used inside of the SF 2D Tools a lot in the Sprite Editor to create new Sprites like Unity's built in Sprite Editor by dragging a Rect over a texture sheet with sprites in it to create a new Sprite asset.