Moving some of the WIP AssetDatabase classes to a new branch.
Making a commit with some of the editor only AssetDatabase classes that will be worked on for the second alpha release.

These will consist of a set of tools to use scriptable objects as simple path definitions to make an organized list of groups from project assets that can be referenced in a single line of code for editor tools, scriptable objects like a set of presets for game logic, and more.

Example use case is having it be used to reference any visual tree uxml asset , style sheet, or image file used for UI Elements in a single line of code.

This might be split into a different package or put into the utility package. Not sure yet.
We can put it into the Utility package and have a specific release branch for including or excluding it for those that don't want bloat in their project, but want to use the SF Utilities.