Made the start of a static utility class to keep track  of common USS classes names in the CommonUSS.uss styles.
The SFCommonStyleClasses static utility file can be used to easily get the string representative values of the uss styles inside of the CommonStyle sheet.

This will be used to easily set combo uss classes as well. For example if you need to set a element as a two column row, but the element is not an SFRow element you can use the TwoColumnRow value here and pass it as a parameter when adding new classes to elements.