Added an Editor only version of SFDataGroup and SFAssetDatabase.
Note both added classes inherit from the runtime versions of them. So if both runtime and editor can use the same code we can just add it to the runtime class versions instead of needing it in both of them.
Added:

- Added a version of both SFDataGroup and SFAssetDatabase that works for only editor.
This is to allow adding editor specific code for SFDataGroups that need it.

Examples:
1. File path data read/write
2. Assetdatabase class function calling.
3. Certain editor only serialization commands. This is for SFDataGroups being used in custom editor tools.

The goal of point three is to make it where we can serialize asset data in any type of database without needing to create a scriptable object per asset.
Think of making a yaml or save file using AES encryption that encrypts the JSON of saved scriptable objects in a single file instead of a asset per entry in the database. When loading/saving the database we could just load that one file. Might want to do a file per database and not a single giant ass file.

This does the following.

1. No more extra folders and assets in the project to search around for.
2. A lot cleaner GitHub repo due to not needing to store hundreds of scriptable objects.