Added an extension method called AbsoluteRectValues function for Rects.
Added:
- Extension method AbsoluteRectValues to make sure Rect's don't get negative values.
This can happen for example if you are making a Rect based on drag inputs. You could start the drag on the right and drag leftward making the width negative.
Useful for certain editor inputs commands. For example the SF 2D Toolkit uses it in the SFSpriteEditor for dragging selecting sprites.