Added two CallerFilePath methods to easily locate the actual asset file path that is calling this.
Added:
- GetCallerPathWithoutExtension to just get the file path to an object in unity without it's file type extension.
- GetCallerPathWithNewExtension to just get the file path to an object in unity while declaring a new file type extension extension.

Example usages
These are used to allow creating things like a SFDataGroup asset and have it do a filtered search for assets in the same folder or subfolder if the user checks that option.

SFDataGroup assets will auto search for Unity assets in the same folder of the declared data type for it's filter, if it is set, and add it automatically to the SFDataGroup category.

Basically you have a folder for Background music with subfolders of certain regions of them game to have their own background music.
Drop a SFDataGroup in the root folder for the audio files and all of them will be added to the SFDataGroup category without the user needing to manually set them in an inspector.